WIP
The logic of Pool::get() is obviously broken, perhaps trying to be too
clever about it.

The really unclear/tricky part of it all is that the emitter wants to
push the newly emitted particle onto its list of particles, but that
list may already include that particular particle, since it may have
been recycled...

Worse, if I wanted all emitters to share from the same pool, then
particles might get sequestered out of another emitter, so it needs to
be removed!

A conundrum.