implement per-emitter pool
This approach also does away with particles removing themselves from
their emitter once they're out of view. This means that particles
continue to be simulated when out of view, but not drawn. Performance is
kept in check by virtue of the limited number of particles that can be
in existence, thanks to the pool.