fix tab bug
Pressing tab focused an element in the panel, moving the viewport of the
body element so the panel was in view. Because the body has hidden
overflow, it's not possible to scroll back. The panel *appears* to be
permanently open.

The fix is to actually remove the panels from the render tree after
hiding them, so they can't receive tab focus.